Edit Site: Move template switcher to the header.
