+++ packages/edit-site/src/components/header/index.js
@@ -1,14 +1,36 @@
 /**
  * WordPress dependencies
  */
+import { useCallback } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
+import { useEditorContext } from '../editor';
+import TemplateSwitcher from '../template-switcher';
 import SaveButton from '../save-button';
 
 export default function Header() {
+	const { settings, setSettings } = useEditorContext();
+	const setActiveTemplateId = useCallback(
+		( newTemplateId ) =>
+			setSettings( ( prevSettings ) => ( {
+				...prevSettings,
+				templateId: newTemplateId,
+				templateType: 'wp_template',
+			} ) ),
+		[]
+	);
+	const setActiveTemplatePartId = useCallback(
+		( newTemplatePartId ) =>
+			setSettings( ( prevSettings ) => ( {
+				...prevSettings,
+				templateId: newTemplatePartId,
+				templateType: 'wp_template_part',
+			} ) ),
+		[]
+	);
 	return (
 		<div
 			className="edit-site-header"
@@ -20,6 +42,14 @@ export default function Header() {
 				{ __( 'Site Editor' ) } { __( '(beta)' ) }
 			</h1>
 			<div className="edit-site-header__actions">
+				<TemplateSwitcher
+					ids={ settings.templateIds }
+					templatePartIds={ settings.templatePartIds }
+					activeId={ settings.templateId }
+					isTemplatePart={ settings.templateType === 'wp_template_part' }
+					onActiveIdChange={ setActiveTemplateId }
+					onActiveTemplatePartIdChange={ setActiveTemplatePartId }
+				/>
 				<SaveButton />
 			</div>
 		</div>
