+++ packages/edit-site/src/components/header/style.scss
@@ -33,5 +33,7 @@
 }
 
 .edit-site-header__actions {
+	align-items: center;
+	display: flex;
 	padding: 0 20px;
 }
